requirements.txt
anytree==2.4.3
certifi==2018.11.29
dask[bag]==1.1.0
dj-database-url==0.5.0
Django==2.1.11
django-cors-headers==3.1.1
django-extensions==2.2.6
django-filter==2.2.0
django-letsencrypt==3.0.1
django-oidc-provider==0.7.0
django-treebeard==4.3
django-user-accounts==2.1.0
django-webpack-loader==0.6.0
elasticsearch==6.3.1
google-auth==1.6.2
google-cloud-pubsub==0.39.1
graphene-django==2.6.0
gunicorn==19.9.0
honcho==1.0.1
lxml==4.3.0
MyCapytain==2.0.9
pinax-eventlog==2.0.3
pinax-theme-bootstrap==8.0.1
pinax-webanalytics==4.0.2
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
raven==6.10.0
regex==2018.11.22
requests==2.22.0
six==1.12.0
whitenoise==4.1.2
wrapt==1.11.1
